Move le files to a more logical location
And rename BatchRoundDataEntry --> BatchRoundTallyEntry for clarity
