--- client/src/components/JurisdictionAdmin/BatchRoundDataEntry/index.ts
@@ -1,3 +1,0 @@
-import BatchRoundDataEntry from './BatchRoundDataEntry'
-
-export default BatchRoundDataEntry
