+++ client/src/components/JurisdictionAdmin/BatchRoundSteps/BatchRoundTallyEntry/index.ts
@@ -1,0 +1,3 @@
+import BatchRoundTallyEntry from './BatchRoundTallyEntry'
+
+export default BatchRoundTallyEntry
