+++ client/src/components/JurisdictionAdmin/BatchRoundSteps/EnterTalliesStep.tsx
@@ -3,7 +3,7 @@ import { toast } from 'react-toastify'
 import { Button } from '@blueprintjs/core'
 import { IJurisdiction } from '../../UserContext'
 import { IRound } from '../../AuditAdmin/useRoundsAuditAdmin'
-import BatchRoundDataEntry from '../BatchRoundDataEntry'
+import BatchRoundTallyEntry from './BatchRoundTallyEntry'
 import { useBatches, useFinalizeBatchResults } from '../useBatchResults'
 import { useConfirm, Confirm } from '../../Atoms/Confirm'
 import { StepPanel, StepActions } from '../../Atoms/Steps'
@@ -55,7 +55,7 @@ const EnterTalliesStep: React.FC<IEnterTalliesStepProps> = ({
   return (
     <>
       <StepPanel>
-        <BatchRoundDataEntry
+        <BatchRoundTallyEntry
           electionId={jurisdiction.election.id}
           jurisdictionId={jurisdiction.id}
           roundId={round.id}
