+++ client/src/components/TallyEntryUser/TallyEntryScreen.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { H2 } from '@blueprintjs/core'
 import { Inner } from '../Atoms/Wrapper'
-import BatchRoundDataEntry from '../JurisdictionAdmin/BatchRoundDataEntry'
+import BatchRoundTallyEntry from '../JurisdictionAdmin/BatchRoundSteps/BatchRoundTallyEntry'
 
 interface ITallyEntryScreenProps {
   electionId: string
@@ -17,7 +17,7 @@ const TallyEntryScreen: React.FC<ITallyEntryScreenProps> = ({
   return (
     <Inner flexDirection="column" withTopPadding>
       <H2>Enter Tallies</H2>
-      <BatchRoundDataEntry
+      <BatchRoundTallyEntry
         electionId={electionId}
         jurisdictionId={jurisdictionId}
         roundId={roundId}
